requirements.txt
pyarrow
tqdm
pandas
matplotlib
scikit-learn
#
pytorch
transformers
# datasets 
# evaluate 
# accelerate 
# gluonts 
# ujson
# matplotlib